Move config helpers to common
